bug - type 'int' is not a subtype of type 'double'

+++ packages/firebase_ml_vision/lib/src/text_recognizer.dart
@@ -97,7 +97,7 @@ abstract class TextContainer {
                 data['height'],
               )
             : null,
-        confidence = data['confidence'],
+        confidence = data['confidence'].toDouble(),
         cornerPoints = List<Offset>.unmodifiable(
             data['points'].map<Offset>((dynamic point) => Offset(
                   point[0],
